Command Refactor
Commands did not have a sufficient way to communicate information back to callers, specifically if those callers were not clients (through chat).

Now commands take a ChatMessage as a parameter, and can build a response to the caller therein, which frees them from the constraints of communicating through the chat system and will enable front-end integration.