World Milestone
Major progress has been made on the WorldContainer implementation (and surronding/supporting classes/functionality).

Have gotten to the point where -- with a test generator -- I can request a block from the world, and the WorldContainer will attempt to load that column (which fails properly since the column was never saved), generate and populate that column, fetch the block, and return it/pass it to an asynchronous callback.

Many implementation details have changed, including the locking mechanism: Exclusive locking is used to manage writes, while reads may be performed at any time.  Lock management is transparent unless you wish to perform a "transaction" -- a large sequence of atomic reads and writes against a world which is guaranteed not to change.

Both synchronous and asynchronous paradigms are supported equally and will continue to be going forward.