World Update -- Maintenance
Upon installation the world module will now begin a maintenance loop.  At an interval which may be configured (but which defaults to once every 5 minutes) all columns will be scanned, saved (if they are "dirty"), and unloaded (if they have no attached players and their interest count is zero).

Factored the code for saving columns out of the maintenance function, and changed the way in which columns are saved: They are copied onto the stack and saved from there, which drastically cuts down on the amount of time a lock is held on the column to save it (since the lock is only held long enough for a quick copy, rather than long enough for a write to the backing store).

Implemented better/more verbose error handling in the process function.

Implemented logging in the set_seed function.

Added code to load WorldContainer settings from the backing store upon installation.

Made various tweaks to the ColumnContainer class to support new/improved/changed functionality.

Columns will no longer be "dirty" after simply being loaded from the backing store.