WorldContainer Interest
Callers may now express interest in WorldContainer columns, causing those columns to be loaded/populated/generated as applicable and to remain loaded until interest is released.

This is necessary so that the spawn area may be generated when the server starts up, rather than when the first player logs in (potentially incurring a long delay).