Time
Added a module to send 0x04 packets (i.e. time updates).

The module also tracks the age of the world, and current time, and allows from those the current time of day, the number of ticks since the last midnight, and the phase of the moon to be derived.

Also supports setting the time, as well as adding to it and subtracting from it.

Also added a /time command, which prints out the current time, phase of the moon, and time of day.