Tick-Based Scheduling
The TimeModule class now provides facilities for scheduling tasks based on ticks.  A task may be run every tick, or a certain number of ticks after it is scheduled.  Moreover these tasks may be fully asynchronous (the tick loop continues whether they finish or not) or only partially asynchronous (they run asynchronously, leveraging the full power of threading and the thread pool, but the next tick will not occur until they finish).

Accordingly the TimeModule class now includes facilities for specifying a tolerance, beyond which a tick is considered "too long" and will cause a warning to be logged.