Set Block
Added the ability to set blocks -- both synchronously and asynchronously -- to the WorldContainer class.

Changed the way the world lock is handled during population.  Previously population consisted of one large transaction, however populators may not require this kind of transactionality, and can start and end their own transactions.  Accordingly, and in the interest of not holding up other operations on the world, the populator no longer starts/ends its own transaction.

Fixed a formatting issue in generator.cpp