Encryption Fix
Fixed two major issues with encryption which were causing frequent client disconnects.

Access to encryptors was not synchronized, meaning that it was possible for two threads to simultaneously attempt an encryption operation, corrupting the encryptor's state.

Between generating an encrypted buffer and sending that buffer there was no synchronization, meaning that it was possible for one thread to obtain an encrypted buffer, be preempted, have a second thread obtain an encrypted buffer and then send it, and then have the first thread resume and send, throwing the client and server's encryption states out of sync since the encrypted buffers would not arrive and be decrypted in the order they were encrypted.