Client Atomic Overhaul
Overhauled the way Client atomic actions were implemented, providing a much greater degree of flexibility.

Previously certain overloads allowed certain operations to be performed atomically in some order.  This limited the atomic operations that could be performed based on the specific overloads that were provided, and also led to a massive amount of code duplication.

Now atomic operations against the Client class are implemented through a variadic template, which can take any number of actions and executes them in whatever order they are passed in.  This gives control over the order of operations, as well as the operations that are performed, directly to the caller, with infinite flexibility.