Time Vanilla Interoperability
Updated the Time class to correct an issue with vanilla Minecraft interoperability:

A Minecraft world begins at 0 ticks (both in age and in time), which translates to 6AM rather than 12AM (as I had previously believed).

Therefore, the client and server were out of sync -- when the client thought it was 6AM, the server thought it was midnight.

This has been corrected by simply adding 6000 to the time ticks before using the value, which translates them between the vanilla Minecraft reality, and a more mathematically convenient model.