Server Factory
Added a static Get method to the Server class, which enables it to be used (by new code) as though it were managed by the singleton-/factory-model used by newer modules/classes.

This is a band-aid to allow both means of access until the code base can be refactored to use exclusively the new model.