Concurrency Manager
Added a class to mcpp.dll which allows thread pools to be managed in a much finer way than simply throwing tasks at them and waiting on the results.

The ConcurrencyManager class takes a number when it is constructed and wraps a thread pool.  Tasks may then be dispatched through the manager.  The manager will ensure that no more than the specified maximum number of tasks that it manages are in the thread pool at any one time.

This allows code that performs a large number of long-running tasks to conveniently enqueue them without saturating the thread pool and starving other tasks.