Singleton/Factory Pattern Refactor
Refactored the entire server code base to use a singleton/factory pattern for module object acquisition/use, which will result in a better, safer, and easier to understand/maintain module loading/unloading paradigm.

The only significant change beyond the changes necessary to accomodate the new pattern (most of which are name changes), is the addition of the "OnShutdown" event to the Server class.  Since modules can no longer guarantee that they will be unloaded after modules which depend on them, if they allow other modules to hook into them, they'll have to handle this event and clear out all such handlers.