MCPP|Brand Handling
Added a module to handle the built-in plugin message channel "MC|Brand", which communicates the "brand" of servers/clients back-and-forth.

This information is sent to clients, received from them, and stored in a way that mods may consume.

Also added a module so operators can view the "brand" of all connected clients.