Command Overhaul
Overhauled the Command class so that it's now possible for commands to be issued independent of an associated client.

This enables other modules to issue commands when a client may not be associated with that command, and allows for front-end command integration.

Added an interface -- CommandInterpreter -- which allows command interpreters to hook themselves into the Server class, for consumption by front-ends.