Front-End Fixes/Improvements
Fixed an arithmetic error in the degenerate case (window height of zero).

The worker thread which maintains the console in the interactive front-end now polls 30 times per second so that the output data is kept appropriately sized (Windows/conhost.exe does not seem to push window resizes to applications in any way, only screen buffer resizes, which are different).