Configuration Commands
Added /set, which allows options in the settings store to be set/unset either in-game or through the interactive front-end (manually inserting/deleting rows in the database no longer necessary).

Added /get, which allows options in the settings store to be retrieved and displayed either in-game or through the interactive front-end.  May consider adding the ability to enumerate all settings (and therefore list them all alongside their values) at some point in the future.