Octave Fix
Fixed the Octave function in noise.hpp so that it no longer takes the arguments to forward through to the generator by universal reference.

As the generator is possibly invoked multiple times by a single invocation of Octave, the arguments to the generator cannot simply be forwarded through, as this might move them, meaning they may not be present in their original form when they're needed next.