Default Generator - Rivers
Rounded out the rivers implementation.  Rivers now "blend" seamlessly into the ocean through the continental shelf.  Rivers now have a depth chosen by the "height" of the world -- in "higher" areas rivers will be shallower, in "lower" areas rivers will be deeper.

Also changed the type of depth-related parameters to an integer type rather than a floating point type.  This makes much more sense given the meaning of the value -- a fractional depth does not make sense in the context of Minecraft.

Removed the "river_depth" parameter as it is no longer needed.