World Improvements
Vastly improved the performance and design of the World class (and supporting classes) by adding the WorldHandle class.

The WorldHandle class is used to read from/write to the world according a pair of strategies -- a write strategy (which dictates whether the writes shall be "dirty" or transactonal) and an access strategy (which dictates how the WorldHandle/World classes should load/generate/populate columns in response to reads and/or writes).

This enables several important changes/optimizations:

The elimination of quite a few overloads -- and many prospective overloads -- by encapsulating and abstracting the different access strategies into a convenient handle.

Reducing the number of column look ups that have to be performed by caching the last column accessed inside the handle.  Assuming subsequent accesses are in the same column (a likely scenario) this eliminates the need for a mutex acquire/release and unordered_map lookup for each and every read or write operation beyond the first.

Population information is encapsulated in the WorldHandle which is passed around by populating threads.  This eliminates the need for mutex acquisitions and releases, and lookups in a hash table of thread IDs to determine whether the current thread is populating.