GCC 4.8.1
Moved to compiling the server with GCC 4.8.1 on Windows.

Accordingly I can now take advantage of the [[noreturn]] standard C++ attribute rather than the GCC-specific __attribute__((noreturn)).

I was hoping GCC 4.8.1 would fix the bug preventing me from deducing the minimum necessary alignment for Variant's internal byte buffer (GCC rejects the constexpr value in the template as not being a constant for the purposes of alignas) but apparently that fix isn't in 4.8.1.  Maybe 4.9.