Add GetBinary Overload
Added a GetBinary overload to the DataProvider interface, and the MySQLDataProvider implementation thereof, which allows the retrieval of an entry from the binary store without an associated pointer/buffer -- one will be created for you.

This allows binary data to be loaded where the size thereof is not known in advance -- a use case that did not exist previously.