Version, Compatibility, and Server Information Overhaul
Overhauled the way that information about the server software itself is supplied, such as compatible protocol/vanilla minecraft version, the name of the software itself, the version of the software, when it was built, and which compiler built it.

Added a method to nicely format versions given a triplet of version numbers (major/minor version and patch number).