Variant Assign/Construct from Convertible Types
If a Variant is assigned/constructed with a type that is not contained within the Variant (or the same type as the Variant, i.e. copy/move construct/assign), the Variant will attempt to find the first type in its template parameter list to which the type-in-question can be converted.

This allows for much more natural syntax, especially where strings and certain integer/floating point types are involved.