Variant Alignment
Found a workaround to a GCC bug I was experiencing: GCC wouldn't allow me to use a constexpr integer to define the alignment of the byte buffer in Variant, so I was just using the alignment of std::max_align_t.

However, writing a slightly more complicated helper template which, rather than deducing the numerical alignment of the most aligned type, deduced the most aligned type itself, rectified this problem.

So now Variants won't be unnecessarily over-aligned, which might save a few bytes here and there.