Client Logging and Receive Fix
Fixed an oversight which could cause decrypted bytes to accumulate and never be parsed -- the Receive method was returning straight away when a buffer with no bytes was passed in, which could happen regularly if a buffer of ciphertext contained two or more packets.

Fixed irregularities in the protocol analysis logging of raw bytes for encrypted clients.

Fixed irregularities in the way that the number of parsed bytes was reported.