World Update
Updated the world module to 1.7.

Rewrote the ToChunkData and GetUnload methods of the ColumnContainer class to leverage the new and improved packet serializer.  While these both seem to work in a basic sense, there appears to be some issues with the new implementations (I'm observing issues with transmitting light values, but that could be originating elsewhere).

Also did some basic/general cleanup.

Removed dependencies on common.hpp.