Makefile Improvement
Made an improvement to Makefile which greatly decreases build times when iteratively building/testing only certain components.

Previously all makefiles would be dragged in and rebuilt everytime the Makefile was invoked, which meant that as you were modifying/testing a certain header file, the makefiles for all files which depended on that header file (many of which were irrelevant at the moment and not being rebuilt) would be rebuilt.

Now only makefiles for those files which are actually being built will be rebuilt, which cuts build times when iteratively modifying/testing a component which includes modifying a header file more than in half.