Network Improvements
Reworked the API for the network stack to make it more general.

The network stack is no longer attached to sockets when it's created.  It may be told to listen on sockets at any time by calling Listen, which returns a ListeningSocket smart pointer.  If Shutdown is called on this the network stack will stop listening.

The network stack may also be leveraged to form outgoing connections using the Connect method.

This allows the threads allocated for networking to be reused by any modules that have to perform custom network operations, reducing the number of worker threads that have to be created for that purpose, thereby reducing context switches, keeping worker threads "warm", and increasing performance.

POSIX is not supported yet.