Authentication API
Refactored the minecraft.net authentication module so that it's now an API and a module which uses that API.

This allows for all server code to share a single Yggdrasil::Client instance, which allows for the HTTPHandler's worker thread to be used by many modules, which decreases the number of threads and thereby the number of possible context switches.

It also hides some relatively ugly code behind a nice abstraction.