Command Line Front-End Improvements
Broke the command line front-end out into a command line wrapper library and a front-end which leverages that library.

This makes the code for that particular front-end substantially cleaner, while allowing other similar front-ends (such as a planned remote command line front-end) to leverage the functionality already implemented for it.