Data Provider Settings Helper
Added a helper member function template to the DataProvider abstract class which allows settings to be retrieved parsed to certain types.  Alternately, if the setting does not exist, or could not be parsed to the type-in-question, a default will be returned.

This will allow for a tremendous reduction in boilerplate code when fetching settings.