Chat Fix
In 1.7.2 Mojang seems to have changed the way that newline characters are rendered by the client.

Previously -- 1.6.4 -- the client would actually render a line break, allowing multi-line chat messages to be sent as one message.

In 1.7.2 this behaviour has been changed.  The client now renders "LF" in a box, and does not move to a newline.

Therefore I had to implement a workaround which transparently splits multi-line messages into multiple packets, and then sends them atomically.