HTTP Handler Error Message Capture
Tweaked the HTTP Handler internals so that the cURL error string associated with a failure would be captured whenever status code 0 -- which represents a failure, rather than an HTTP status code -- would be passed through to the completion handler.

This will allow for better debug logging and therefore better problem diagnosis.