Permissions
Replaced the old operator system with a new permissions system, which allows fine-grained control over permissions on a per user or per group basis.

This permissions system has a module which implements four commands -- /grant, /revoke, /addgroup, and /removegroup -- which allow for the manipulation of permissions, as well as an information provider -- /info permissions -- which allows permission tables to be viewed.