Serializer
Added a generic binary serialization framework which depends on the DataProvider interface to get and save data.

This framework is templated, which means that adding specializations to serialize/deserialize new types is easy/trivial.

As the server grows, this will allow mods to save and load complex data easily and using tested, reusable, generic code.