Connection Handler Information
Fixed bugs of omission which were causing the ConnectionHandler class' Disconnected and Accepted statistics to be forever reported as zero.

Added a getter function to the Server class to retrieve the server's ConnectionHandler.

Added Listening, Connected, and Workers statistics to the ConnectionHandler class.

Added an information provider for the handler class (/info handler).

Fixed a typo in documentation of the ConnectionHandlerInfo class in network.hpp.