Settings Overhaul
Overhauled the manner in which settings are extracted from a DataProvider.  Renamed the pure virtual GetSetting function to RetrieveSetting.  This is now merely a dummy worker function which is called by the new, templated GetSetting function, and may become protected at some point in the future.

The new, templated GetSetting function behaves similarly to the GetSetting function which takes a default and coalesces to that if the setting-in-question either does not exist, or cannot be converted to the type-in-question, except the new GetSetting function returns a nulled Nullable<T> if the setting-in-question does not exist, or could not be converted.

Overhauled the SettingConverter class and specializations for better separation of responsibilities (previously the coalescing functionality was baked into SettingConverter).