Connection Fixes
Fixed a possible deadlock in the Connection class, which would result if the connection was terminated while a thread was both holding a reference to the connection (thus keeping it alive) and waiting on a pending send through that same connection.

Fixed a possible access violation if a Connection object outlives its associated ConnectionHandler.