Login/Logout Broadcast Update
Updated the module which broadcasts login/logout events through global chat.

The code depended on the ClientState enumeration, which was removed during the update to 1.7.x protocol.

More importantly, however, the code for this module was a total mess, having been written before a lot of the methods/features/etc. of MCPP which allow cleaner/more concise modules to be written were implemented/added/discovered.