Player List Module Update
Updated the module which keeps the TAB player lists in the client up-to-date.

The new method for insuring all clients have a consistent view of the players that are on-line (and that players appear on-/off-line immediately) is far superior to the old one.

Added the corresponding packet definition -- Packets::Play::Clientbound::PlayerListItem -- to packet.hpp.