Ban Module
Added a ban module.  Together the blacklist module and this new ban module implement and improve on the functionality of the old (and long removed) ban module.

The ban module allows usernames to be banned.  When a username is banned, any connected users with that username will be disconnected, and incoming connections from people with that username will be blocked (by disconnecting them in response to the OnLogin event).  A ban may have a username (the person who enacted the ban) and reason associated with it.

Implemented two commands -- /ban and /unban -- which allow the banlist to be manipulated, and an info provider -- /info bans -- which allows the banlist to be viewed.  The ban module itself provides an API through which these functions -- and more -- are exposed for use in any further modules.

Added a template which allows nullable types to be serialized/deserialized to serializer.hpp.