Whitelist
Added a whitelist system.

Can be toggled on/off either by a configuration setting (in which case it will be on or off when the server starts) or while the server is running.

Players on the whitelist will be allowed to login, whereas other players will be blocked from doing so.

Also added a command -- /whitelist -- for manipulating the system, and an information provider -- /info whitelist -- for displaying whether the whitelist is enabled, and which players are on the whitelist.