Time Update
Updated the time/tick manager module.  In addition to making it compatible with the changes made to the entire project in the wake of 1.7.x I significantly rewrote it, so it's much more readable/maintainable.

Also added a forgotten #pragma once, and a forgotten doxygen \file directive in time.hpp.

The /time command is now an information provider -- /info time -- and provides significantly more information.