IPAddress Hash Change
Realized that the way that I had implemented std::hash for IPAddress objects was ridiculous.  An IPAddress is either a UInt128 -- if it's IPv6 -- or a UInt32 -- if it's not.  I should've just been using the std::hash specializations for those integer types.

In changing the std::hash implementation over, realized that while GCC supplies a 128-bit integer type, it doesn't supply std::hash specializations for it, so I implemented those as well.