Promise
Added a Promise<T> template, which encapsulates the idea of a promised, future return value, for use with asynchronous functions going forward.

In contrast to the built-in std::future and std::promise, provides a Then function, which attaches a function to received the promised value.  The Then function, in turn, returns yet another Promise<T>, for the result of the specified function.