libcurl Wrapper Rewrite
This has been coming for a while (since the original wrapper was written) as the original was a total mess.

The new wrapper:

Uses the SocketPair class to be more responsive.  Due to the lack -- previously -- of a good way to wake up the worker while it was sleeping on select (since Windows doesn't have socketpair natively and libcurl doesn't allow for IOCP to be integrated easily), the worker just set an arbitrary timeout and was completely unresponsive while waiting for socket events.  Newly-added easy handles were liable to sit untouched, and a shutdown was liable to block unnecessarily.  This has all been rectified.

Handles all HTTP requests it's capable of handling (limited to GET, HEAD, and POST at the moment) through a single method.  This method takes an HTTPRequest object, which carries all information about the request.  This allows for much cleaner code to be built on top of the handler, as code that issues a variety of HTTP requests can easily build shared code to handle/dispatch them.

Uses promises.  Making a request returns to the caller a promise of an HTTPResponse object, which contains all data the handler parses from the server.  Unlike the old handler this leaves the choice of how to proceed -- block or continuation -- to the caller, increasing caller flexibility/ease of use.  The fact that promises pass exceptions through also means that the handler communicates failures in the proper C++ manner -- by throwing -- which makes its API much less confusing/easier to consume.

Is much cleaner.  The code separates responsibilities much better, and is much cleaner/more readable as a result.

Properly handles alternative character encodings in the response.  Previously the handler made the assumption that all text was UTF-8.  The handler is now capable of properly parsing the HTTP Content-Type header and determining which decoder to use based on that (falling back to UTF-8).  Decoding is now done lazily, so rather than decoding the response as it comes in, raw bytes are returned in the HTTPResponse object, which the caller may lazily have decoded (by a method on the HTTPResponse class).  This increases caller flexibility, as the caller has access to the raw bytes of the response (in case -- for example -- something other than text is being returned).