Yggdrasil Rewrite
Completely rewrote the Yggdrasil client to use the newly-rewritten HTTPHandler and to use Promise<T> for asynchronous return values.

The new client has a much cleaner API, especially with respect to error reporting.  Rather than passing through variants whose engaged type specified whether the request succeeded or failed, the promise exception throwing mechanism is used to throw the appropriate exception if the request fails.  This makes the API much easier to consume, and is also in line with good C++ style/form.