Scope Guard
Added a ScopeGuard class, which simply wraps a functor and calls it when the object goes out of scope.

Added two overloads of a helper function template -- AtExit -- which deduce and wrap a functor, possibly with arguments forwarded through and bound into it (one of them is variadic, one of them is not -- the non-variadic one avoids the overhead of invoking std::bind).