Server Fixes
The server's panic-related methods were not properly flagged noreturn.  Adding this attribute caused the compiler to reveal that on some code paths some of the panic-related methods actually did return -- not intended at all.  Accordingly this has been fixed so these methods are now guaranteed not to return -- if a user-supplied panic callback happens to return, std::abort is unconditionally invoked.

The Connection::Shutdown method on the POSIX network stack is not noexcept.  Corrected a piece of code in the receive loop to reflect this -- the call to Client::Disconnect is now properly guarded by a try/catch with a call to panic if it happens to throw.