Resource Pool
Implemented a resource pool as a generic class -- Pool<T>.

The pool allows exclusive resources to be acquired and released in a thread safe manner.  If the pool does not have enough of a certain resource -- either because there's nothing currently in the pool or because all other resources are currently acquired -- it will call a user-supplied callback to obtain another resource.  It will do this until it reaches a configurable maximum (which may be unlimited), at which point incoming threads will block until a resource becomes available.